Moved files from filter.js to seed.js
