+++ index.html
@@ -43,7 +43,7 @@
             </div>
             <div class="card-content">
                 <h3>Craftsmen Charm</h3>
-                <p class="prices">Price: $3,000</p>
+                <p class="prices">Price: ${{randomPrice}}</p>
                 <p class="rooms">Bedroom: 3</p>
                 <p class="bathrooms">Bathroom: 3</p>
                 <p>City: San Jose</p>
