+++ package.json
@@ -18,6 +18,7 @@
     "express": "^4.19.2",
     "express-handlebars": "^7.1.3",
     "express-session": "^1.18.0",
+    "handlebars": "^4.7.8",
     "pg": "^8.12.0",
     "sequelize": "^6.37.3"
   },
