+++ views/partials/card-details.handlebars
@@ -1,13 +1,13 @@
 <div class="card">
-    <div class="img-container">
+    <div class="img-container"">
         <img src="image">
     </div>
-    <div class="card-content">
-        <h3>title</h3>
+    <div class="card-content" id="card-template">
+        <h3>{{title}}</h3>
         <p>Price: ${{price}}</p>
         <p>Bedroom: {{bedroom}}</p>
         <p>Bathroom: {{bathroom}}</p>
-        <p>City: city</p>
-        <p>State: state</p>
+        <p>City: {{city}}</p>
+        <p>State: {{state}}</p>
     </div>
 </div>
